pyproject.toml
[project]
name = "slack-ai-chat-prototype"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "httpx>=0.27.0",
  "pydantic>=2.8.0",
  "pydantic-settings>=2.4.0",
  "presidio-analyzer>=2.2.0",
  "presidio-anonymizer>=2.2.0",
  "google-cloud-firestore>=2.16.0",
  "google-genai>=0.5.0",
  "slack-sdk>=3.33.0",
  "pytest>=8.3.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

